+++ src/lib/randomAfanumeric.ts
@@ -1,0 +1,14 @@
+import { randomBytes } from 'crypto';
+
+/**
+ * Generate a random afanumeric string of given lenght
+ */
+const generateRandomAlphanumeric = (num: number) => {
+    return randomBytes(Math.ceil(num * (3 / 4)))
+        .toString('base64')
+        .slice(0, num)
+        .replace(/\+/g, '0')
+        .replace(/\//g, '0');
+};
+
+export default generateRandomAlphanumeric;